requirements.txt
Django
Pillow
tensorflow